libhdfesse: use fixed config in tests
We do not load auto config in the integration tests, as it can be
dangerous if the tests are run on a wrong host.  However, we test the
auto config against the standard config to test config loading.  It's
ok if this test fails when run on a wrong host.